입찰(롤 구분으로 업체,일반)하기 경매등록 경매조회 판매종료(입찰자 없으면 없는대로) 판매 자동종료 1분주기 스케줄러

+++ src/main/java/web/mvc/dto/BidDTO.java
@@ -24,7 +24,7 @@ public class BidDTO {
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss", timezone = "Asia/Seoul")
     private LocalDateTime insertDate;
 
-    private Integer auctionSeq;
+    private Long auctionSeq;
 
-    private Integer userSeq;
+    private Long userSeq;
 }
